packages/django/pyproject.toml
[tool.poetry]
name = "django-quanttide-projects"
version = "0.1.0-beta.1"
description = "量潮项目管理Django工具箱"
authors = ["QuantTide Inc. <opensource@quanttide.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "django_quanttide_projects"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = ">=3.1"
djangorestframework = ">=3.0"
django-quanttide = {git = "https://github.com/quanttide/django-quanttide.git"}

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.7"
tox = "^4.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
